refactor: move reducers to their own files
