+++ apps/src/code-studio/pd/workshop_dashboard/WorkshopFormTemplate/index.tsx
@@ -17,6 +17,8 @@ import {workshopLabel} from '../utils/workshopLabel';
 import {DATE_FORMAT, TIME_FORMAT} from '../workshopConstants';
 
 import {generateNewSession} from './components/SessionsEditor';
+import {sessionsReducer} from './reducers/sessionsReducer';
+import {workshopReducer} from './reducers/workshopReducer';
 import AdditionalInfo from './sections/AdditionalInfo';
 import Basics from './sections/Basics';
 import EmailsReminders from './sections/EmailsReminders';
@@ -27,10 +29,8 @@ import Schedule from './sections/Schedule';
 import {
   FieldConfig,
   Session,
-  SessionAction,
   SessionFormState,
   Workshop,
-  WorkshopAction,
   WorkshopFormState,
   WorkshopFormTemplateProps,
 } from './types';
@@ -84,89 +84,6 @@ export const sessionDataToState = (
     format: session.session_format ?? 'in_person',
     sameAsPrevious: false,
   }));
-
-export const workshopReducer = (
-  state: WorkshopFormState,
-  action: WorkshopAction
-): WorkshopFormState => {
-  switch (action.type) {
-    case 'UPDATE_WORKSHOP':
-      return {...state, ...action.payload};
-    case 'ADD_GRADE': {
-      const newGrades = state.grades.concat(action.payload);
-      newGrades.sort((a, b) => {
-        // sort 'K' to beginning
-        if (a === 'K') return -1;
-        if (b === 'K') return 1;
-        // sort 'Other' to end
-        if (a === 'Other') return 1;
-        if (b === 'Other') return -1;
-        const numA = Number(a);
-        const numB = Number(b);
-        if (isNaN(numA) || isNaN(numB)) return 0;
-        return numA - numB;
-      });
-      return {...state, grades: newGrades};
-    }
-    case 'REMOVE_GRADE':
-      return {
-        ...state,
-        grades: state.grades.filter(grade => grade !== action.payload),
-      };
-    case 'ADD_COURSE_OFFERING':
-      return {
-        ...state,
-        courseOfferings: [...state.courseOfferings, action.payload],
-      };
-    case 'REMOVE_COURSE_OFFERING':
-      return {
-        ...state,
-        courseOfferings: state.courseOfferings.filter(
-          offering => offering !== action.payload
-        ),
-      };
-    case 'SET_COURSE_OFFERINGS':
-      return {
-        ...state,
-        courseOfferings: action.payload,
-      };
-    case 'SET_WORKSHOP':
-      return action.payload;
-    default:
-      return state;
-  }
-};
-
-export const sessionsReducer = (
-  state: SessionFormState[],
-  action: SessionAction
-): SessionFormState[] => {
-  switch (action.type) {
-    case 'ADD_SESSION':
-      return state.concat(generateNewSession(state[state.length - 1]));
-
-    case 'UPDATE_SESSION':
-      return state.map(session =>
-        session.id === action.id ? {...session, ...action.payload} : session
-      );
-
-    case 'UPDATE_SESSION_SAME_AS_PREVIOUS':
-      return state.map((session, i) =>
-        session.id === action.id
-          ? {...session, ...(state[i - 1] ?? {}), sameAsPrevious: true}
-          : session
-      );
-
-    case 'DELETE_SESSION':
-      return state.filter(session => session.id !== action.id);
-
-    case 'SET_SESSIONS':
-      return action.payload;
-
-    default:
-      return state;
-  }
-};
 
 export const WorkshopFormTemplate: FC<WorkshopFormTemplateProps> = ({
   config,
@@ -355,7 +272,6 @@ export const WorkshopFormTemplate: FC<WorkshopFormTemplateProps> = ({
     }),
     [workshopErrors.registrationLink, workshopErrors.hidden]
   );
-
 
   const sectionProps = useMemo(
     () => ({
