+++ apps/src/code-studio/pd/workshop_dashboard/WorkshopFormTemplate/reducers/sessionsReducer.ts
@@ -1,0 +1,33 @@
+import {generateNewSession} from '../components/SessionsEditor';
+import {SessionFormState, SessionAction} from '../types';
+
+export const sessionsReducer = (
+  state: SessionFormState[],
+  action: SessionAction
+): SessionFormState[] => {
+  switch (action.type) {
+    case 'ADD_SESSION':
+      return state.concat(generateNewSession(state[state.length - 1]));
+
+    case 'UPDATE_SESSION':
+      return state.map(session =>
+        session.id === action.id ? {...session, ...action.payload} : session
+      );
+
+    case 'UPDATE_SESSION_SAME_AS_PREVIOUS':
+      return state.map((session, i) =>
+        session.id === action.id
+          ? {...session, ...(state[i - 1] ?? {}), sameAsPrevious: true}
+          : session
+      );
+
+    case 'DELETE_SESSION':
+      return state.filter(session => session.id !== action.id);
+
+    case 'SET_SESSIONS':
+      return action.payload;
+
+    default:
+      return state;
+  }
+};
